requirements/train_requirements.txt
torch
torchvision
numpy
pandas
Pillow
scikit-learn
wandb
tqdm
datasets
cloudml-hypertune
kaggle
google-cloud-secret-manager